pyproject.toml
[tool.poetry]
name = "openai-fine-tuning"
version = "0.1.0"
description = ""
authors = ["Mathew Wang"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
openai = "^1.40.3"
python-dotenv = "^1.0.1"
pandas = "^2.2.2"
tiktoken = "^0.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
